requirements.txt
discord.py[voice]
yandex-music
python-dotenv
PyNaCl